Make COM declaration tests independent of built JS artifacts
Share the isolated COM runtime declaration fixture between overload and sink tests. Keep strict TypeScript and negative assertions while removing the coverage-time dependency on bindings/js/dist.

Copilot-Session: 9ef7fa55-eb2a-4360-bd17-05feae50fadb
